Upgrade the Docker image to Fedora 31
DockerHub keeps a cached image of fedora:latest when the latest
release was Fedora 28.

When running this script with --oem-string on Fedora 28 we get:

  qemu-system-x86_64: -smbios type=11,value=: Don't know how to build fields for SMBIOS type 11

To be able to use the SMBIOS type 11, we need a QEMU more recent
that the 2.11.2 that is shipped in Fedora 28.

We can use this option with Fedora 29 (shipped with QEMU 3.0.1),
but this release is also obsolete. Use the recent Fedora 31 as a
base for testing, with QEMU 4.1.1.

Signed-off-by: Philippe Mathieu-Daudé <philmd@redhat.com>