Use philmd docker image based on more recent Fedora (31)
Current image hosted in the puiterwijk/ Docker Hub namespace is
based on Fedora 28, which is too old to use the QEMU feature to
build SMBIOS type 11 data.
Upgrade the image and push it to my namespace (philmd/) to test
this unbreak the Circle CI issue.
Patrick Uiterwijk has write access to my namespace. We can keep
this image here, or we can revert this commit later.

Signed-off-by: Philippe Mathieu-Daudé <philmd@redhat.com>